requirements.txt
adan==0.0.2
chumpy==0.70
data==0.4
easymocap.egg==info
einops==0.4.1
filterpy==1.4.5
imageio==2.27.0
imgaug==0.4.0
ipdb==0.13.13
lap==0.4.0
matplotlib==3.5.2
mediapipe==0.8.10.1
mmcv_full==1.7.0
mmdet==2.26.0
mmdet.egg==info
modelscope==1.4.2
numpy==1.22.3
numpy==1.23.2
numpy==1.23.5
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python==4.7.0.72
opencv_python_headless==4.6.0.66
Pillow==9.5.0
Pillow==9.5.0
pyrender==0.1.45
pytorch3d==0.7.2
PyYAML==6.0
PyYAML==6.0
scikit_image==0.19.3
scipy==1.7.3
scipy==1.9.1
sdf_pytorch==0.0.1
sdf_pytorch.egg==info
setuptools==62.1.0
setuptools==65.6.3
tabulate==0.9.0
tensorboardX==2.4
tensorboardX==2.6
timm==0.5.4
timm==0.6.12
torch==1.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.64.0
tqdm==4.65.0
trimesh==3.9.24
triton==2.0.0.dev20221030
visdom==0.2.4
visualize==0.5.1
yacs==0.1.8
~mcv_full==1.5.1
